+++ lib/rubex/ast/statement/begin_block.rb
@@ -1,0 +1,14 @@
+module Rubex
+  module AST
+    module Statement
+      module BeginBlock
+        class Base < Statement::Base
+          def initialize(statements, location)
+            @statements = statements
+            super(location)
+          end
+        end
+      end
+    end
+  end
+end
